Add picture language.
The picture language is broken in mit-scheme-9.2 because it requires the
student band which does not seem to compile and also uses
floating-vector-* instead of flo:vector-* procedures.

This commit ports over the library and removes some MIT-Scheme 6001 band
specific configurations that were causing failure.

The language must be used from inside the MIT-Scheme REPL. To use it,
fire up the REPL and run `(load "lib/picture/psgo.scm")`.

Then start to play `(paint g1 fovnder)` (yes, that's a v not a u).